# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZalvadoraTest API Documentation",
        "description": "API para el sistema de gestión de suscripciones de la prueba técnica.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión de usuario",
                "operationId": "bae68d1c55329ecb6e41f48285840c8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@empresa.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso con token"
                    },
                    "401": {
                        "description": "Credenciales no válidas"
                    }
                }
            }
        },
        "/api/v1/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Listar todas las empresas",
                "operationId": "707a45d5190cbbe1b43f69f6e5d0131d",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Registrar una nueva empresa",
                "operationId": "8c4a1d582268449c242009ff20c2dfe7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterCompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado"
                    }
                }
            }
        },
        "/api/v1/companies/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Obtener una empresa por ID",
                "operationId": "f0f1a3d86918a593a0aaa37a0c9fb1c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Actualizar una empresa",
                "operationId": "e1521de8d65f0ab70f079174789b8b9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Borrar una empresa",
                "operationId": "4b2b739cc0dc94bdc5feebb74e4e9914",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sin contenido"
                    }
                }
            }
        },
        "/api/v1/companies/{company}/change-plan": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Cambiar el plan de una empresa",
                "operationId": "fd06c47400cc155293d12c0a8f5efefe",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeCompanyPlanRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Sin contenido"
                    }
                }
            }
        },
        "/api/v1/companies/{company}/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Listar usuarios de una empresa",
                "operationId": "541185c593aca1d78995e156b9a1a69c",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Añadir un nuevo usuario a una empresa",
                "operationId": "025cc737315adff61691fd5c711be8c9",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado"
                    },
                    "422": {
                        "description": "Límite de usuarios excedido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/companies/{company}/users/{user}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Actualizar un usuario específico",
                "operationId": "6a745ccc97e790d9084b761bafab7ab3",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Borrar un usuario específico",
                "operationId": "8274792de266d4a70da3b3b8b2915867",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuario eliminado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/plans": {
            "get": {
                "tags": [
                    "Plans"
                ],
                "summary": "List all plans",
                "operationId": "68377f3912c8e992d9c9bd723f220357",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Plans"
                ],
                "summary": "Create a new plan",
                "operationId": "be2f3d755929ad01dcec2cf20bf1a008",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePlanRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plan created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/plans/{id}": {
            "get": {
                "tags": [
                    "Plans"
                ],
                "summary": "Get a single plan by ID",
                "operationId": "fe151362f2d33475ea9d5f1f7a130baf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Plans"
                ],
                "summary": "Update a plan",
                "operationId": "3c513db4a32589f6f6ab5a04eaea2398",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePlanRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plan updated successfully"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Plans"
                ],
                "summary": "Delete a plan",
                "operationId": "a9a7d6150000ff4144d4f0ce8662740e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Plan deleted successfully"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ChangeCompanyPlanRequest": {
                "description": "* @OA\\Schema(required={\"new_plan_id\"})",
                "required": [
                    "new_plan_id"
                ],
                "properties": {
                    "new_plan_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "RegisterCompanyRequest": {
                "required": [
                    "name",
                    "email",
                    "plan_id"
                ],
                "properties": {
                    "plan_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "StorePlanRequest": {
                "required": [
                    "name",
                    "price",
                    "user_limit"
                ],
                "properties": {
                    "features": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Soporte 24/7"
                        ]
                    }
                },
                "type": "object"
            },
            "StoreUserRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "UpdateCompanyRequest": {
                "title": "Update Company Request",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "nuevo@email.com"
                    }
                },
                "type": "object"
            },
            "UpdatePlanRequest": {
                "title": "Update Plan Request",
                "description": "Datos para la actualización de un plan.",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Plan Básico Plus"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 25.5
                    },
                    "user_limit": {
                        "type": "integer",
                        "example": 15
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Soporte mejorado"
                        ]
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "title": "Update User Request",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.updated@empresa.com"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Autenticación con token Bearer (Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Companies",
            "description": "Companies"
        },
        {
            "name": "Users",
            "description": "Users"
        },
        {
            "name": "Plans",
            "description": "Plans"
        }
    ]
}